refact(api): merge knowledge_models.py into schemas_knowledge.py (#5996)
Eliminates dedicated model file in favour of domain schema file.
Updated 10 importers across knowledge_* endpoint files.
