+++ autobot-backend/api/knowledge_verification.py
@@ -23,7 +23,7 @@ from datetime import datetime, timezone
 
 from fastapi import APIRouter, Depends, HTTPException, Path, Query, Request
 
-from api.knowledge_models import (
+from api.schemas_knowledge import (
     PendingSourceResponse,
     VerificationConfig,
     VerificationRequest,
